[10.0][MIG][REF] Splitted files based in purchase and purchase line objects.
